+++ CHANGES.md
@@ -7,7 +7,7 @@ Features ✨:
 Improvements 🙌:
  - You can now join room through permalink and within room directory search
  - Add long click gesture to copy userId, user display name, room name, room topic and room alias (#1774)
- - Do not propose to verify session if there is only one session (#1901)
+ - Do not propose to verify session if there is only one session and 4S is not configured (#1901)
 
 Bugfix 🐛:
  - Display name not shown under Settings/General (#1926)